+++ docs/next-i18next.config.cjs
@@ -5,6 +5,6 @@ module.exports = {
   i18n: {
     defaultLocale: 'en',
     locales: ['en', 'zh'],
-    localePath: path.resolve('./public/locales'),
   },
+  localePath: path.resolve('./public/locales'),
 };