+++ packages/components/src/utils/forwardRef.tsx
@@ -14,5 +14,5 @@ export default function forwardRef<Props extends object, Component extends As>(
     }
   >
 ) {
-  return React.forwardRef(component) as unknown as ComponentWithAs<Component, Props>;
+  return React.forwardRef(component as any) as unknown as ComponentWithAs<Component, Props>;
 }